Bugfix when parsing data generators arguments
The problem came when the command / argument index gets high in the input data
editing. For example, an editing command 8 as well as an editing command 880. In
this case, the arguments to command 880 where delivered to command 8 (matching
the common "*::ARG8"). In order to prevent this partial matching, the pattern
has been extended with the necessary final "::".