In some cases, the line number exclusions were not working.
This happened when the same file was read multiple times (for example,
when trying to read a time period bigger than the buffer size). A micro
optimization in how to compare the latest line restriction with the
current line number would be stuck to the end of file line number,
preventing the application of the previous time restrictions.