Better solution for issue 919
The previous "hacky" way was actually not working properly (the tests did not
pass). I was happy to come with a smart solution based on regex and negative
lookbehind (with such as regex: (.*)(?<!\\)(?:#|;)(.*) ) but lookbehind is not
yet implemented by c++ compilers :-( (it seems that it was specified in 2018 and
it takes time until it gets implemented in compilers and the said compilers get
into the users' hands! So we now have a simpler, low tech solution, but it
works...

By the way, I've just learned about lookbehind expressions when looking at ways
to fix the test! :-)