Restructuring of libacdd
It is now cleaner and more flexible. It is possible to define
environment variables for any ACDD configuration key as a way to provide
more relevant defaults, system-wide. There is also finer control on the
dependencies between acdd fields (and it will be possible to implement
even more if necessary).