Big restructuring in the dataGenerators.
This is required in order to allow some generators (like TauCLD) to look
before/after the requested data point in order to be able to process the
requested data point (for example in order to interpolate the cloudiness
over the night). The code is TauCLDGenerator is messy (lots of commented
out code), but it works as before (until the upcoming improvements).