Removed the getXXXPart() methods as they are redundant with existing calls.
A new getValuesRegex() has been added for completeness and some documentation
added for clarity. The name of the interpolation1D section is back as a const
string in Meteo1DInterpolator as a way to keep everything in one place for each
particular implementation (knowing that this is a public member and as such can
be read by other classes if necessary).